Issue #138: Moving AnalyzerResultFuture to DataCleaner-api project to fix testware cyclic dependency with engine-core.
