+++ api/src/main/java/org/datacleaner/api/AnalyzerResultFuture.java
@@ -17,7 +17,7 @@
  * 51 Franklin Street, Fifth Floor
  * Boston, MA  02110-1301  USA
  */
-package org.datacleaner.result;
+package org.datacleaner.api;
 
 import java.io.IOException;
 import java.io.ObjectOutputStream;
@@ -28,8 +28,6 @@ import java.util.concurrent.CountDownLatch;
 import org.apache.metamodel.util.HasName;
 import org.apache.metamodel.util.Ref;
 import org.apache.metamodel.util.SharedExecutorService;
-import org.datacleaner.api.Analyzer;
-import org.datacleaner.api.AnalyzerResult;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
